Added the before_request so that all calls must have a valid token. Next is to get all the rest tests to pass again.
