+++ app/status/healthcheck.py
@@ -4,7 +4,7 @@ from flask import Blueprint
 status = Blueprint('status', __name__)
 
 
-@status.route('/_status')
+@status.route('/_status', methods=['GET', 'POST'])
 def show_status():
     return jsonify(
         status="ok",
